Added very basic Frontend UI
Implemented the basic structure for the first version of our "game."
I added a mainpage page that could be our main menu. Currently it has a find match button. This button takes the user to the "game" page. The user must click 50 times and is trying to beat the opponents score of 10 seconds.
Next we need to add the actual match against another player.